requirements.txt
fastapi
uvicorn
motor
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
email-validator
